requirements.txt
apache-airflow
pandas
sqlalchemy
python-dotenv
psycopg2-binary
requests